moleculesコンポーネント分割 #19
